requirements.txt
# 基础依赖
requests==2.31.0
PyYAML==6.0.1
tqdm==4.66.1
python-dotenv==1.0.0

# 文献爬取
arxiv==2.1.0
scholarly==1.7.11

# PDF处理
PyMuPDF>=1.24.10
pdfplumber==0.10.3

# 数据库
# sqlite3 is included in Python standard library
psycopg2-binary==2.9.9  # 用于PostgreSQL（可选）
SQLAlchemy==2.0.23      # 用于ORM（可选）

# 报告生成
Jinja2==3.1.2
matplotlib==3.8.2
wordcloud==1.9.3

# LLM相关
langchain==0.1.13

# 测试
pytest==7.4.3

# 其他工具
unidecode==1.3.6  # 用于文本处理
beautifulsoup4==4.12.2  # 用于HTML解析
lxml==4.9.3  # 用于XML解析
